feat: Move getLatestEthBlockhash to auth-browser
