+++ apps/html/index.html
@@ -1,4 +1,4 @@
-<!-- (DO NOT EDIT!) (HTML) THIS FILE IS AUTOMATICALLY GENERATED FROM tools/scripts/gen-html.mjs Tue, 14 Nov 2023 16:08:37 GMT  -->
+<!-- (DO NOT EDIT!) (HTML) THIS FILE IS AUTOMATICALLY GENERATED FROM tools/scripts/gen-html.mjs Mon, 20 Nov 2023 12:23:55 GMT  -->
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -62,7 +62,7 @@ pre {
 </style> 
 </head>
 <body>
-    (HTML) THIS FILE IS AUTOMATICALLY GENERATED FROM tools/scripts/gen-html.mjs Tue, 14 Nov 2023 16:08:37 GMT
+    (HTML) THIS FILE IS AUTOMATICALLY GENERATED FROM tools/scripts/gen-html.mjs Mon, 20 Nov 2023 12:23:55 GMT
 
     <!-- ==================== ALL EXPORTED VANILLA LIBRARIES ==================== -->
     <script src="dist/packages/access-control-conditions-vanilla/access-control-conditions.js"></script>
