+++ apps/react/src/app/app.tsx
@@ -1802,7 +1802,7 @@ pre {
                     `,
                 }}
             />
-            (REACT) THIS FILE IS AUTOMATICALLY GENERATED FROM tools/scripts/gen-react.mjs Tue, 14 Nov 2023 16:08:38 GMT
+            (REACT) THIS FILE IS AUTOMATICALLY GENERATED FROM tools/scripts/gen-react.mjs Mon, 20 Nov 2023 12:23:55 GMT
             <div id="root"></div>
              <pre><code id="result"></code></pre>
         </>
