+++ packages/constants/src/lib/utils/utils.ts
@@ -1,5 +1,3 @@
-import { ethers } from 'ethers';
-import { LIT_EVM_CHAINS } from '../constants/constants';
 import { EITHER_TYPE } from '../enums';
 import { IEither, ILitError } from '../interfaces/i-errors';
 
@@ -29,28 +27,4 @@ export function ERight<T>(result: T): IEither<T> {
     type: EITHER_TYPE.SUCCESS,
     result,
   };
-}
-
-export async function getLatestEthBlockhash(): Promise<string> {
-  // Not using the first rpc as it always returns the same nonce. Seems like the RPC just returns a dummy value
-  for (let i = 1; i < LIT_EVM_CHAINS['ethereum'].rpcUrls.length; i++) {
-    const provider = new ethers.providers.JsonRpcProvider(
-      LIT_EVM_CHAINS['ethereum'].rpcUrls[i]
-    );
-
-    try {
-      console.log('Fetching the latest Eth blockhash from RPC- ', i);
-      const latestEthBlockhash = (await provider.getBlock("latest")).hash;
-      console.log('latestEthBlockhash- ', latestEthBlockhash);
-
-      return latestEthBlockhash;
-    } catch (e: any) {
-      console.error(
-        'Error getting the latest Eth blockhash- ' + i + ', trying again: ',
-        e
-      );
-    }
-  }
-
-  throw new Error('Unable to get the latestBlockhash from any RPCs');
 }
